Add a const file
