+++ lib/const.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+const double iconSize = 36;
+const double counterNumbersSize = 52;
+const double timerNumbersSize = 36;
+const double resultNumbersSize = 44;
+const double noResultFontSize = 32;
+const buttonColor = Color(0xFF5C5862);
